优化代码结构 & 增加单节点限速配置支持
Took 2 hours 34 minutes

+++ folib-storage/folib-storage-api/src/main/java/com/veadan/folib/services/impl/ConfigurationManagementServiceImpl.java
@@ -154,7 +154,7 @@ public class ConfigurationManagementServiceImpl
     }
 
     @Override
-    public void setKbps(Long kbps) throws IOException {
+    public void setKbps(Integer kbps) throws IOException {
         modifyInLock(configuration -> configuration.setKbps(kbps));
     }
