+++ folib-storage/folib-storage-core/src/main/java/com/veadan/folib/services/ConfigurationManagementService.java
@@ -32,7 +32,7 @@ public interface ConfigurationManagementService {
 
     void setPort(int port) throws IOException;
     
-    void setKbps(Long kbps) throws IOException;
+    void setKbps(Integer kbps) throws IOException;
     
     void setSliceMbSize(Long sliceMbSize) throws IOException;
