+++ folib-web-forms/src/main/java/com/veadan/folib/forms/configuration/ClusterDispatchNodeForm.java
@@ -23,5 +23,5 @@ public class ClusterDispatchNodeForm {
     /**
      * 节点传输速率（KB/s）
      */
-    private Long kbps;
+    private Integer kbps;
 }
