+++ folib-web-forms/src/main/java/com/veadan/folib/forms/configuration/ServerSettingsForm.java
@@ -35,7 +35,7 @@ public class ServerSettingsForm
     /**
      * 节点传输速率（KB/s）
      */
-    private Long kbps;
+    private Integer kbps;
     /**
      * 节点传输切片大小（MB）
      */
@@ -105,11 +105,11 @@ public class ServerSettingsForm
         this.port = port;
     }
 
-    public Long getKbps() {
+    public Integer getKbps() {
         return kbps;
     }
 
-    public void setKbps(Long kbps) {
+    public void setKbps(Integer kbps) {
         this.kbps = kbps;
     }
